Starting to refactor unit tests
