+++ src/test/java/frc/robot/state/sequencer/SequenceStateMachineTest.java
@@ -51,12 +51,6 @@ public class SequenceStateMachineTest {
         runAlgaeScoringSequence(sequence);
     }
 
-    @Test
-    void testFloorPickupAlgae(){
-        Sequence sequence = getSequence(Action.INTAKE, Level.L1, GamePiece.ALGAE);
-        runPickupAlageFromFloor(sequence);
-    }
-
     /*
      * COMMON METHODS FOR RUNNING SEQUENCES
      */
@@ -166,44 +160,6 @@ public class SequenceStateMachineTest {
         // transition back to home
         callback.setInput(SequenceInput.ELEVATOR_DONE);
         callback.setInput(SequenceInput.ARM_DONE);
-        assertEquals(SequenceState.HOME, stateMachine.getCurrentState());
-    }
-
-    private void runPickupAlageFromFloor(Sequence sequence){
-        // mock the subsystems
-        ElevatorSubsystem mockedElevatorSubsystem = mock(ElevatorSubsystem.class);
-        ArmSubsystem mockedArmSubsystem = mock(ArmSubsystem.class);
-        HandIntakeSubsystem mockHandIntakeSubsystem = mock(HandIntakeSubsystem.class);
-        HandClamperSubsystem mockHandClamperSubsystem = mock(HandClamperSubsystem.class);
-
-        SequenceStateMachine stateMachine = new SequenceStateMachine(mockedElevatorSubsystem, mockedArmSubsystem, mockHandClamperSubsystem, mockHandIntakeSubsystem);
-        Positions positions = SequenceFactory.getPositions(sequence);
-        StateMachineCallback callback = stateMachine.getSubsystemCallback();
-        stateMachine.setSequence(sequence);
-
-        // verify initial state
-        assertTrue(stateMachine.isReady());
-        assertEquals(SequenceState.HOME, stateMachine.getCurrentState());
-
-        // transition to moving arm
-        stateMachine.setInput(SequenceInput.BEGIN);
-        verify(mockedArmSubsystem).moveArm(positions.armForwardPosition, callback);
-        assertEquals(SequenceState.MOVING_ARM_FORWARD, stateMachine.getCurrentState());
-
-        // transition to intaking algae
-        callback.setInput(SequenceInput.ARM_DONE);
-        verify(mockHandClamperSubsystem).open(positions.handClamperPosition);
-        verify(mockHandIntakeSubsystem).intake(HandConstants.intakeVelocity, callback);
-        assertEquals(SequenceState.INTAKING, stateMachine.getCurrentState());
-
-        // transition to moving arm home
-        callback.setInput(SequenceInput.DETECTED_PIECE);
-        //verify(mockedArmSubsystem).moveArm(ArmConstants.armHomePosition, callback);    // error here "org.mockito.exceptions.verification.TooManyActualInvocations at SequencerStateMachineTest.java:211"
-        assertEquals(SequenceState.FINISHING, stateMachine.getCurrentState());
-
-        // transition to home
-        callback.setInput(SequenceInput.ARM_DONE);
-        //Check that safty check ran and returned true?
         assertEquals(SequenceState.HOME, stateMachine.getCurrentState());
     }
 }
